Document modules proposal and distinguish between source, tokens, and AST version; update docs with other refinements
Modules will be defined with `module-info.sy`, similar to
`module-info.java` from Java 9. The removed `internal` accessibility
modifier proposal has been restored, but for module-wide rather than
package-wide accessibility.  This is to reflect the use of Rust's new
`pub(crate)` form and C#'s `internal` in regards to .NET assemblies.

Revert the change to the Sydoc proposal by making it start with `/**`
and end with `*/` again. While it looks better for small documentation
due to its being inside the thing it documents, it makes it harder to
see whether a large multi-page comment block is a SyDoc or just a
large comment.

Formally ban scope shadowing and document an example of the impact in
the example source.

Tweak the `continue` keyword proposal to remove the parentheses. It
isn't a function or method call and is clearly a special language
built-in, so parentheses make no sense. Not having them is more
consistent with `if` and `for`. Resolve the subsequent parsing
difficulty of continuing to an explicit label by putting the label
before the `continue` instead of after it. This is still easy to
visually parse as only two variants of this will exist: `LABEL: for` and
LABEL: continue`.

Update the documentation code snippet to match the example source file.
A Cargo build step should be established to automatically insert it in
rather than copying changes across manually.

Document a unified aliasing system and remove class and interface
assignments. This is to make aliasing more consistent and to come up
with a passable syntax for aliasing functions without relying on
explicit argument forwarding.

Document the new Go 1.11-inspired module system and versioning proposal.

Open the language up to the possibility of compilation again, but
emphasise that interpretation will still be the first step. This came
about due to my realisation that a runtime could be exposed to both the
interpreter directly but also to the compiler by exposing demangled
symbols, compiling it as a separate crate, and then statically linking
it with the result.

Document the difference between items and expressions better. Emphasise
the expressions are contained within items but that items cannot be
contained within expressions.

Document the use of type inference better, and that type inference
across function or method block boundaries is not allowed and nor is any
form of structural typing.

Add token stream versions and AST versions to the language. This should
allow macros to not break backwards compatibility with the existing
language versioning.

Expand the documentation's contents page by linking to head Details
subitem.

Point those wanting implementation documentation from the documents to
the RustDoc in the `main.rs` entrypoint as a starting point.