Remove need for parser excursions and ambiguous parsing by overhauling lambda syntax
Ambiguous parsing, along with recursive lexing for string
interpolations, were the two biggest problems to be solved before work
on finishing the parser was committed to. They would have added too much
complexity to lexing and parsing for too little benefit.

Cutting back string interpolations to only work with identifiers rather
than arbitrary expressions was an easy decision since it would also
discourage cryptic Sylan one-liners in strings. Lambdas were a harder
trade-off though, striking a balance between concision, familiarity with
Java and C#, supporting statically-typed destructuring in parameter
lists, and not requiring ambiguous parsing rules.

The new lambda syntax _starts_ with an opening brace, like so:

    arrayList.filter({ x -> x < 2 })

This allows lambdas to be parsed unambiguously, but means opening braces
can no longer be used as "grouped expressions" in code like this:

    var x = {
        println("Hello, world!")
        42
    }

Now the `do` keyword replaces this, similar to Perl's do blocks:

    var x = do {
        println("Hello, world!")
        42
    }

`do` was previously used for context blocks, a.k.a. monadic do notation,
so a new keyword `with` replaces it. `do` needn't be a keyword, it could
actually be a prelude function that uses syntactical sugar detailed
below.

Forcing braces is more verbose than the previous form of just using an
arrow for single expressions, so syntactical sugar has been adopted from
Ruby and Kotlin:

* If a lambda has no parameters, the whole parameter list and arrow can
  be dropped:

    Task({ println("Hello, world!") })

* If a lambda is the sole argument to a function or method, the
  parentheses can be dropped:

    Task { println("Hello, world!") }

* If the lambda takes a single parameter, the parameter list and arrow
  can be dropped and the `it` keyword can be used in place of the first
  parameter:

    arrayList.filter { it < 2 }

Unlike Ruby and Swift but like Kotlin, Sylan won't allow passing lambdas
outside of calling parentheses when other arguments are also passed:

    times(2) {
        println("Hello, world!")
    }

This is because an opening brace would be interpreted as a new
expression yielding a lambda, the `times` invocation would be treated as
a seperate expression, which would then fail to type check due to the
wrong amount of arguments being passed.

Besides, the real advantage of that construct is that it looks a bit
more like a language built-in, but it actually won't in Sylan since
other control flow keywords don't require parentheses, unlike Java and
C#.

To summarise, I spent hours rediscovering why Kotlin went with the
lambda syntax they chose, since Sylan has ended up essentially doing a
1-to-1 clone of it. Nicely done, Kotlin team.