Ok, allow a `-` prefix on numeric literals but still disallow as unary
This is a compromise between keeping the lexing and parsing simple and
consistent and keeping the language similar to Java and C#. Having both
unary and binary operators is fine, except when one operator can do
both. That creates silly situations in which just whitespace alone is
enough to cause ambiguities, which is worsened by the fact that Sylan,
unlike Java, does not have explicit line terminations.

Case in point: what does this mean?

    2
    -4

Evaluate 2 and then evaluate -4, or evaluate `2 - 4`, yielding `-2`?
As we can lex the whole of `-4` as a token, this can be resolved as the
first, but what about this?

    var x = { 4 }
    2
    -(x())

This is just confusing. So, the compromise is to allow the `-` prefix as
a binary operator, as a prefix to a whole number token, but _not_ as an
arbitrary unary operator.

Some background context to this problem:
https://mail.haskell.org/pipermail/haskell-cafe/2006-September/017941.html