Merge contextual and non-contextual scopes, and make contextual bindings expressions rather than special binding-like items
A contextual binding is just syntactical sugar for passing a lambda to a
monadic bind function, so developers don't expect them to have the same
coreferencing abilities of normal bindings.

The ability for `var` bindings to reference each other without one being
predeclared is the only reason why bindings are treated specially in
scopes, a feature that allows frictionless corecursion among other
things.