Migrate proposal from Python-like keyword arguments to Swift-like parameter labels
This fixes the problem of usages of the same function being inconsistent
depending on which positional arguments the developer decided to use as keyword
arguments.

Now it's left up to the definer of a function or method, ensuring it is used
consistently.

Keyword arguments are used by default to encourages readable functions.
Adopting a "positional-only" feature from Python 3.8 would be less elegant than
instead adopting Swift's parameter labels.

This has another nice side-effect; the syntactical sugar for passing a variable
as a keyword argument with the same name, the dot-prefix, can be removed from
the language. Now parameters are either keyword or positional with no
ambiguity, so this can be worked out at compile time without additional syntax.