Put SyDocs after declarations again and reintroduce constants
SyDocs going afterwards was originally dropped because of confusion of position
with constants, but putting them straight after the constant works fine.

Constants were reintroduced because this then created another ambiguity about
whether SyDocs should be allowed on `var` or not, depending on whether it's in
a main package. This is one ambiguity of a long list, so let's bite the bullet
and just distinguish module-level constants from local bindings.

This also resolves another open problem: should module imports be allowed to
trigger side-effects? Erlang doesn't think so, and I'm increasingly taking its
views. This can be now fixed by just mandating compile-time constants for
constants defined with `final`.