Start taking capability-based security seriously in proposal
Propose that the runtime send an extern task to the main task on startup, which it can then optionally send to subtasks to transfer ambient authority.

When certain subsystems are invoked, they'll expect to see it in the mailbox.

This basically reimplementing the old dynamic scoping proposal for ambient authority with messaging across processes.

Clarify that tasks are totally opaque in Sylan IL proposal.

Define a few utility functions for managing authority, such as `dropAmbientAuthority` and `spawnTrusted`.